julia free mousepos trace added

+++ fractol.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   fractol.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fwlpe <fwlpe@student.42.fr>                +#+  +:+       +#+        */
+/*   By: cdenys-a <cdenys-a@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/28 13:34:44 by fwlpe             #+#    #+#             */
-/*   Updated: 2019/04/09 18:27:48 by fwlpe            ###   ########.fr       */
+/*   Updated: 2019/04/10 16:43:48 by cdenys-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -69,7 +69,7 @@ typedef struct	s_fctl
 	double		*im;
 	t_cam		cam;
 	int			pxs;
-	int			*cl_store;
+	double		*cl_store;
 	t_cl		cl;
 	char		*f_name;
 }				t_fctl;
